Remove NFont project and tidy up code
