+++ BloomFramework.sln
@@ -4,16 +4,8 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 15.0.27703.2018
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BloomFramework", "BloomFramework\BloomFramework.vcxproj", "{C640C874-D67E-4B31-AF02-378FE5C4D3A3}"
-	ProjectSection(ProjectDependencies) = postProject
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6} = {03AB7317-F8EC-49C1-8917-13A94A63B2B6}
-	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Test Bench", "Test Bench\Test Bench.vcxproj", "{69BCBA61-1EF9-4E33-92FB-DC40D2F2319C}"
-	ProjectSection(ProjectDependencies) = postProject
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6} = {03AB7317-F8EC-49C1-8917-13A94A63B2B6}
-	EndProjectSection
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "nfont", "nfont\nfont.vcxproj", "{03AB7317-F8EC-49C1-8917-13A94A63B2B6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -39,14 +31,6 @@ Global
 		{69BCBA61-1EF9-4E33-92FB-DC40D2F2319C}.Release|x64.Build.0 = Release|x64
 		{69BCBA61-1EF9-4E33-92FB-DC40D2F2319C}.Release|x86.ActiveCfg = Release|Win32
 		{69BCBA61-1EF9-4E33-92FB-DC40D2F2319C}.Release|x86.Build.0 = Release|Win32
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6}.Debug|x64.ActiveCfg = Debug|x64
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6}.Debug|x64.Build.0 = Debug|x64
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6}.Debug|x86.ActiveCfg = Debug|Win32
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6}.Debug|x86.Build.0 = Debug|Win32
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6}.Release|x64.ActiveCfg = Release|x64
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6}.Release|x64.Build.0 = Release|x64
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6}.Release|x86.ActiveCfg = Release|Win32
-		{03AB7317-F8EC-49C1-8917-13A94A63B2B6}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
