--- nfont/.gitmodules
@@ -1,3 +1,0 @@
-[submodule "SDL_FontCache"]
-	path = SDL_FontCache
-	url = https://github.com/grimfang4/SDL_FontCache.git
