--- nfont/SDL_FontCache/test/test.cbp
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<CodeBlocks_project_file>
-	<FileVersion major="1" minor="6" />
-	<Project>
-		<Option title="test" />
-		<Option pch_mode="2" />
-		<Option compiler="gcc" />
-		<Build>
-			<Target title="Debug">
-				<Option platforms="Windows;" />
-				<Option output="./test" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj/Debug/" />
-				<Option type="1" />
-				<Option compiler="gcc" />
-				<Compiler>
-					<Add option="-Wall" />
-					<Add option="-g" />
-					<Add directory="../externals/SDL2/include/SDL2" />
-					<Add directory="../externals/SDL_ttf/include" />
-					<Add directory=".." />
-				</Compiler>
-				<Linker>
-					<Add library="mingw32" />
-					<Add library="SDL2main" />
-					<Add library="SDL2" />
-					<Add library="SDL2_ttf" />
-					<Add directory="../externals/SDL2/lib_windows" />
-					<Add directory="../externals/SDL_ttf/lib_windows" />
-				</Linker>
-			</Target>
-			<Target title="Debug - SDL_gpu">
-				<Option platforms="Windows;" />
-				<Option output="./test-SDL_gpu" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj/Debug-SDL_gpu/" />
-				<Option type="1" />
-				<Option compiler="gcc" />
-				<Compiler>
-					<Add option="-Wall" />
-					<Add option="-g" />
-					<Add option="-DFC_USE_SDL_GPU" />
-					<Add directory="../externals/SDL2/include/SDL2" />
-					<Add directory="../externals/SDL_ttf/include" />
-					<Add directory=".." />
-					<Add directory="../externals/SDL_gpu/include" />
-				</Compiler>
-				<Linker>
-					<Add library="mingw32" />
-					<Add library="SDL2main" />
-					<Add library="SDL2_gpu" />
-					<Add library="SDL2" />
-					<Add library="SDL2_ttf" />
-					<Add library="OpenGL32" />
-					<Add directory="../externals/SDL2/lib_windows" />
-					<Add directory="../externals/SDL_ttf/lib_windows" />
-					<Add directory="../externals/SDL_gpu/lib_windows" />
-				</Linker>
-			</Target>
-			<Target title="Emscripten">
-				<Option output="test" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj/Emscripten" />
-				<Option type="0" />
-				<Option compiler="emscripten_compiler" />
-				<Compiler>
-					<Add option="`sdl2-config --cflags`" />
-				</Compiler>
-				<Linker>
-					<Add option="`sdl2-config --libs`" />
-				</Linker>
-			</Target>
-		</Build>
-		<Compiler>
-			<Add option="-Wall" />
-			<Add option="-fexceptions" />
-			<Add option="-DNFONT_USE_TTF" />
-		</Compiler>
-		<Unit filename="../SDL_FontCache.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="../SDL_FontCache.h" />
-		<Unit filename="main.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Extensions>
-			<code_completion />
-			<debugger />
-			<envvars />
-		</Extensions>
-	</Project>
-</CodeBlocks_project_file>
