--- nfont/SDL_FontCache/test/utf8_sample.txt
@@ -1,1 +1,0 @@
-﻿Μπορώ να φάω σπασμένα γυαλιά χωρίς να πάθω τίποτα.
