--- nfont/ideas.txt
@@ -1,4 +1,0 @@
-
-
-Load other bitmap font layouts?
-
