--- nfont/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="sdl2.nuget" version="2.0.8.2" targetFramework="native" />
-  <package id="sdl2.nuget.redist" version="2.0.8.2" targetFramework="native" />
-  <package id="sdl2_image.nuget.redist" version="2.0.3.2" targetFramework="native" />
-  <package id="sdl2_ttf.nuget" version="2.0.14.2" targetFramework="native" />
-  <package id="sdl2_ttf.nuget.redist" version="2.0.14.2" targetFramework="native" />
-</packages>
