--- nfont/test/test.cbp
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<CodeBlocks_project_file>
-	<FileVersion major="1" minor="6" />
-	<Project>
-		<Option title="test" />
-		<Option pch_mode="2" />
-		<Option compiler="gcc" />
-		<Build>
-			<Target title="NFont test">
-				<Option platforms="Windows;" />
-				<Option output="./test-NFont" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj/NFont/" />
-				<Option type="1" />
-				<Option compiler="gcc" />
-				<Compiler>
-					<Add option="-g" />
-					<Add option="-DFC_USE_SDL_GPU" />
-					<Add option="-DNFONT_USE_SDL_GPU" />
-					<Add directory="../externals/SDL2/include/SDL2" />
-					<Add directory="../externals/SDL_ttf/include" />
-					<Add directory="../externals/SDL_gpu/include" />
-					<Add directory="../SDL_FontCache" />
-				</Compiler>
-				<Linker>
-					<Add option="-mwindows" />
-					<Add library="mingw32" />
-					<Add library="SDL2main" />
-					<Add library="SDL2_gpu" />
-					<Add library="SDL2" />
-					<Add library="SDL2_ttf" />
-					<Add library="OpenGL32" />
-					<Add directory="../externals/SDL2/lib_windows" />
-					<Add directory="../externals/SDL_ttf/lib_windows" />
-					<Add directory="../externals/SDL_gpu/lib_windows" />
-				</Linker>
-			</Target>
-			<Target title="NFontR test">
-				<Option platforms="Windows;" />
-				<Option output="./test-NFontR" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj/NFontR/" />
-				<Option type="1" />
-				<Option compiler="gcc" />
-				<Compiler>
-					<Add option="-g" />
-					<Add option="-DNFONT_USE_NFONTR" />
-					<Add directory="../externals/SDL2/include/SDL2" />
-					<Add directory="../externals/SDL_ttf/include" />
-					<Add directory="../SDL_FontCache" />
-				</Compiler>
-				<Linker>
-					<Add option="-mwindows" />
-					<Add library="mingw32" />
-					<Add library="SDL2main" />
-					<Add library="SDL2" />
-					<Add library="SDL2_ttf" />
-					<Add directory="../externals/SDL2/lib_windows" />
-					<Add directory="../externals/SDL_ttf/lib_windows" />
-				</Linker>
-			</Target>
-		</Build>
-		<Compiler>
-			<Add option="-Wall" />
-			<Add option="-fexceptions" />
-			<Add option="-DNFONT_USE_TTF" />
-		</Compiler>
-		<Unit filename="../NFont/NFont.cpp" />
-		<Unit filename="../NFont/NFont.h" />
-		<Unit filename="../SDL_FontCache/SDL_FontCache.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="../SDL_FontCache/SDL_FontCache.h" />
-		<Unit filename="main.cpp" />
-		<Extensions>
-			<code_completion />
-			<debugger />
-			<envvars />
-		</Extensions>
-	</Project>
-</CodeBlocks_project_file>
